requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
reportlab>=4.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
fpdf2>=2.7.0
pytest>=8.0.0
scikit-learn>=1.4.0
